+++ src/component/authenticate/Login.js
@@ -1,7 +1,7 @@
 import { useForm } from "react-hook-form";
 import TextField from "@mui/material/TextField";
 import Button from "@mui/material/Button";
-import NavigationBar from "../navgation/NavigationBar";
+import NavigationBar from "../navgation/navigation-bar";
 import { Link } from "react-router-dom";
 import "./login.css";
 import { useNavigate } from "react-router-dom";
