+++ src/component/authenticate/sign-up.js
@@ -1,16 +1,51 @@
 import { useForm } from "react-hook-form";
-import { stateList, countryList, roleList, townCityList } from "../../master/MasterList";
+import {
+  stateList,
+  countryList,
+  roleList,
+  townCityList,
+} from "../master/master-list";
 import TextField from "@mui/material/TextField";
 import MenuItem from "@mui/material/MenuItem";
 import Button from "@mui/material/Button";
 import Stack from "@mui/material/Stack";
 import Box from "@mui/material/Box";
 import { Link } from "react-router-dom";
+import axios from "axios";
+import NavigationBar from "../navgation/navigation-bar";
+import "./sign-up.css";
+import { useEffect } from "react";
+import { useRef } from "react";
+import { useState } from "react";
+import { useNavigate } from "react-router-dom";
 
-import NavigationBar from "../../navgation/NavigationBar";
-import "../../authenticate/signup.css";
+function Signup() {
+  const [hospitalList, setHospitalList] = useState([]);
+  const [isRegister,setIsRegister] = useState(false)
 
-function PatientRegistration() {
+
+  const navigate = useNavigate();
+
+  useEffect(() => {
+    const getHospitalList = async () => {
+      try {
+        const response = await axios.get(
+          "http://localhost:8080/registration/listHospital"
+        );
+        console.log(response.data);
+
+        setHospitalList(response.data);
+
+      } catch (error) {
+        console.error(error);
+      }
+    };
+
+    getHospitalList();
+  }, []);
+
+
+
   const {
     register,
     handleSubmit,
@@ -18,9 +53,45 @@ function PatientRegistration() {
     watch,
   } = useForm();
 
-  const onsubmit = (data) => {
+  const onsubmit = async (data) => {
     console.log(data);
+    const path = "http://localhost:8080/authenticate/signup";
+    const body = {
+      emailID: data.emailId,
+      password: data.password,
+      hospitalName: data.hospital,
+      docID_empID: data.userID,
+      role: data.role,
+      firstName: data.firstName,
+      lastName: data.lastName,
+    };
+    const obj = {
+      method: "POST",
+      headers: { "Content-Type": "application/json" },
+      body: JSON.stringify(body),
+    };
+
+    try {
+      const response = await fetch(path, obj);
+      const res = await response.json();
+      console.log(res);
+      if (res){
+        setIsRegister(true)
+
+        setTimeout(() => {
+          navigate("/login")
+        }, 2000);
+      }
+      else{
+        setIsRegister(false)
+      }
+
+    } catch (error) {
+      console.error(error);
+    }
   };
+
+  console.log(hospitalList);
 
   return (
     <>
@@ -35,9 +106,9 @@ function PatientRegistration() {
             autoComplete="off"
           >
             <form onSubmit={handleSubmit(onsubmit)}>
-            <div className="form-control">
-              <h1 className="login-text">SIGNUP</h1>
-            </div>
+              <div className="form-control">
+                <h1 className="login-text">SIGNUP</h1>
+              </div>
               <div className="sign-up-elements-div">
                 <TextField
                   className="signup-text-field"
@@ -45,15 +116,26 @@ function PatientRegistration() {
                   label="Hospital Name *"
                   placeholder="Apollo Hospital"
                   variant="standard"
+                  select
                   {...register("hospital", {
                     required: {
                       value: true,
                       message: "Hospital Name is required",
                     },
                   })}
-                  error={!!errors.hospital}
-                  helperText={errors?.hospital?.message}
-                />
+                >
+
+                  {hospitalList.length > 0 ? (
+                    hospitalList.map((option) => (
+                      <MenuItem key={option.hospitalName} value={option.hospitalName}>
+                        {option.hospitalName}
+                      </MenuItem>
+                    ))
+                  ) : (
+                    <></>
+                  )}
+
+                </TextField>
               </div>
 
               <Stack spacing={2} direction="row">
@@ -271,7 +353,7 @@ function PatientRegistration() {
                     backgroundColor: "#7EDD6F",
                     justifyContent: "center",
                     paddingLeft: "60px",
-                  paddingRight: "60px",
+                    paddingRight: "60px",
                   }}
                   variant="contained"
                   type="submit"
@@ -281,12 +363,14 @@ function PatientRegistration() {
               </div>
 
               <div className="form-control">
-              <Link className="signup-link" to="/login">
-                {" "}
-                Already a member?lets login!
-              </Link>
-            </div>
+                <Link className="signup-link" to="/login">
+                  {" "}
+                  Already a member?lets login!
+                </Link>
+              </div>
             </form>
+
+            {isRegister && <div>User is registered successfully</div>}
           </Box>
         </div>
       </div>
@@ -294,4 +378,4 @@ function PatientRegistration() {
   );
 }
 
-export default PatientRegistration;
+export default Signup;
