+++ src/component/dashboard/doctor/doctor-dashboard.js
@@ -1,4 +1,4 @@
-import NavigationBar from "../../navgation/NavigationBar";
+import NavigationBar from "../../navgation/navigation-bar";
 import { useEffect } from "react";
 import { useNavigate } from "react-router-dom";
 import { useSelector } from "react-redux";
