+++ src/component/dashboard/nurse/nurse-dashboard.js
@@ -1,4 +1,4 @@
-import NavigationBar from "../../navgation/NavigationBar";
+import NavigationBar from "../../navgation/navigation-bar";
 import { useNavigate } from "react-router-dom";
 import { useSelector } from "react-redux";
 import { useState, useEffect } from "react";
