+++ src/component/hospital-registration/hospital-registration.js
