+++ src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import reportWebVitals from './reportWebVitals';
-import App from './App';
+import reportWebVitals from './report-web-vitals';
+import App from './app';
 import { legacy_createStore as createStore } from "redux";
 import userLoginReducer from './reducers/userReducers';
 import { Provider } from 'react-redux';
